pyproject.toml
[tool.poetry]
name = "eventhub-to-timescale"
version = "0.1.0"
description = ""
authors = ["rob al <mnbf9rca@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "eventhub_to_timescale"}]

[tool.poetry.dependencies]
python = ">3.9,<3.12"
azure-functions = "^1.12.0"
python-dateutil = "^2.8.2"
psycopg = {extras = ["binary"], version = "^3.1.8"}
jsonschema = "^4.17.3"
python-dotenv-vault = "^0.6.3"
bimmer-connected = "^0.14.2"
azure-data-tables = "^12.4.4"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"
black = "^23.10.1"



[tool.poetry.group.test.dependencies]
flake8 = "^6.0.0"
pytest = "^7.4.2"
pytest-cov = "^4.0.0"
coverage = {extras = ["toml"], version = "^7.3.2"}
pytest-mock = "^3.10.0"
mock = "^5.0.1"
jsonschema = "^4.17.3"
pytest-asyncio = "^0.21.1"
flake8-pyproject = "^1.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.coverage.run]
omit = [
    "*/test/*",
    "*/tests/*",
    "*/venv/*",
    "*/.venv/*",
    "*/.tox/*",
    "*/.eggs/*",
    "*/.vscode-remote/*"
]
relative_files = true

[tool.pytest.ini_options]
addopts = "--cov --cov-branch --cov-report term-missing -vv --strict-markers -rfE"
pythonpath = "${workspaceFolder}"

[tool.flake8]
max-line-length = 120
exclude = [
    ".git",
    "__pycache__",
    "docs/source/conf.py",
    "old",
    "build",
    "dist",
    "venv"
]